feapi/pyproject.toml
[tool.poetry]
name = "feapi"
version = "0.1"
description = "Mocking version of backend for Packing service"
authors = ["Mikhail Lazarenko <mikhail.lazarenko@bluedinamyc.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
mongoengine = "^0.24.1"
mongomock = "^4.0.0" 
grpclib = "^0.4.3"
betterproto = { version = "2.0.0b4", allow-prereleases = true}
pyhumps = "^3.7.0"
dnspython = "^2.2.0"
python-multipart = "^0.0.5"
python-keycloak = "^2.5.0"
python-jose = "^3.3.0"
fastapi_keycloak = "^1.0.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = ["feapi", "bluelib"]
profile = "black"

[tool.black]
line-length = 118

[tool.mypy]
ignore_missing_imports = true
exclude = [
    "/api/",
]
disallow_untyped_defs = true

[tool.pylint.'LINE LENGTH']
max-line-length = 118

[tool.pytest.ini_options]
asyncio_mode = "auto"
